requirements.txt
Django==4.1.5
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
psycopg2==2.9.5
django-cors-headers==3.13.0
Pillow

ccxt
django_redis
aiohttp
celery
gunicorn
pandas
numpy
ta
django-environ
openai
